Further set default nodata_out values to numpy types. Fix for numpy >2.0

+++ pysheds/pgrid.py
@@ -1034,8 +1034,8 @@ class Grid(object):
             fdir_0.flat[prop_0 == 0] = fdir_1.flat[prop_0 == 0]
             fdir_1.flat[prop_1 == 0] = fdir_0.flat[prop_1 == 0]
             # Set nodata cells to zero
-            fdir_0[invalid_cells] = 0
-            fdir_1[invalid_cells] = 0
+            fdir_0[invalid_cells] = np.int64(0)
+            fdir_1[invalid_cells] = np.int64(0)
             # Create indexing arrays for convenience
             domain = np.arange(fdir.size, dtype=np.min_scalar_type(fdir.size))
             unique = np.zeros(fdir.size, dtype=np.bool)
@@ -1251,11 +1251,11 @@ class Grid(object):
                     nodata_cells = (fdir == nodata_in)
             invalid_cells = ~np.in1d(fdir.ravel(), dirmap)
             invalid_entries = fdir.flat[invalid_cells]
-            fdir.flat[invalid_cells] = 0
+            fdir.flat[invalid_cells] = np.int64(0)
             # Ensure consistent types
             fdir = fdir.astype(mintype)
             # Set nodata cells to zero
-            fdir[nodata_cells] = 0
+            fdir[nodata_cells] = np.int64(0)
             # Get matching of start and end nodes
             startnodes, endnodes = self._construct_matching(fdir, domain,
                                                             dirmap=dirmap)
@@ -1805,8 +1805,8 @@ class Grid(object):
                 fdir_0.flat[prop_0 == 0] = fdir_1.flat[prop_0 == 0]
                 fdir_1.flat[prop_1 == 0] = fdir_0.flat[prop_1 == 0]
                 # Set nodata cells to zero
-                fdir_0[invalid_cells] = 0
-                fdir_1[invalid_cells] = 0
+                fdir_0[invalid_cells] = np.int64(0)
+                fdir_1[invalid_cells] = np.int64(0)
                 # Create indexing arrays for convenience
                 visited = np.zeros(fdir.size, dtype=np.bool)
                 # nvisited = np.zeros(fdir.size, dtype=int)
